Move to Vercel
See #38 on motivation for this move.
