+++ pyproject.toml
@@ -22,6 +22,9 @@ ipdb = "^0.13.9"
 cryptography = "^39.0.1"
 
 
+[tool.poetry.group.dev.dependencies]
+black = "^23.3.0"
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
