+++ requirements.txt
@@ -1,0 +1,7 @@
+flask-cors==3.0.10 ; python_version >= "3.8" and python_version < "4.0"
+flask-sqlalchemy==3.0.3 ; python_version >= "3.8" and python_version < "4.0"
+flask==2.2.3 ; python_version >= "3.8" and python_version < "4.0"
+python-dotenv==0.21.1 ; python_version >= "3.8" and python_version < "4.0"
+pymysql==1.0.2 ; python_version >= "3.8" and python_version < "4.0"
+certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4"
+sqlalchemy==1.4.46 ; python_version >= "3.8" and python_version < "4.0"
