--- server/mysql.env.example
@@ -1,1 +1,0 @@
-SQLALCHEMY_DATABASE_URI='mysql+pymysql://username:password@localhost:3306/awesome_manim_video'
